осоздала новую ветку от оappeal_comment букез слияния с develop

+++ src/main/java/kg/attractor/xfood/dto/CheckTypeShowDto.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 public class CheckTypeShowDto {
     private Long id;
     private String name;
-    private Long numsOfCriteria;
+    private Integer numsOfCriteria;
     private Integer totalValue;
     private BigDecimal fee;
 }
